+++ packages/sitecore-jss-nextjs/src/middleware/redirects-middleware.test.ts
@@ -1490,6 +1490,67 @@ describe('RedirectsMiddleware', () => {
       expect(finalRes.status).to.equal(expected.status);
     });
 
+    it('should return 301 redirect when queryString is ordered by alphabetic(Netlify feature)', async () => {
+      const setCookies = () => {};
+      const res = createResponse({
+        url: 'http://localhost:3000/found/',
+        status: 301,
+        setCookies,
+      });
+      const nextRedirectStub = sinon.stub(NextResponse, 'redirect').callsFake((url, init) => {
+        const status = typeof init === 'number' ? init : init?.status || 307;
+        return ({
+          url,
+          status,
+          cookies: { set: setCookies },
+          headers: res.headers,
+        } as unknown) as NextResponse;
+      });
+      const req = createRequest({
+        nextUrl: {
+          pathname: '/not-found/',
+          search: '?a=1&w=1',
+          href: 'http://localhost:3000/not-found/?a=1&w=1',
+          locale: 'en',
+          origin: 'http://localhost:3000',
+          clone() {
+            return Object.assign({}, req.nextUrl);
+          },
+        },
+      });
+
+      const { middleware, fetchRedirects, siteResolver } = createMiddleware({
+        pattern: '/not-found?w=1&a=1',
+        target: 'http://localhost:3000/found/',
+        redirectType: REDIRECT_TYPE_301,
+        isQueryStringPreserved: true,
+        locale: 'en',
+      });
+
+      const finalRes = await middleware.getHandler()(req);
+
+      validateDebugLog('redirects middleware start: %o', {
+        hostname: 'foo.net',
+        language: 'en',
+        pathname: '/not-found/',
+      });
+
+      validateEndMessageDebugLog('redirects middleware end in %dms: %o', {
+        headers: {},
+        redirected: undefined,
+        status: 301,
+        url: 'http://localhost:3000/found/',
+      });
+
+      expect(siteResolver.getByHost).to.be.calledWith(hostname);
+      // eslint-disable-next-line no-unused-expressions
+      expect(fetchRedirects.called).to.be.true;
+      expect(finalRes).to.deep.equal(res);
+      expect(finalRes.status).to.equal(res.status);
+
+      nextRedirectStub.restore();
+    });
+
     describe('should redirect to normalized path when nextjs specific "path" query string parameter is provided', () => {
       it('should return 301 redirect', async () => {
         const setCookies = () => {};